+++ diskover-web/CHANGELOG.md
@@ -1,12 +1,17 @@
 # Diskover-web v2 Community Edition Change Log
 
-# [2.0-rc.4] - 2022-02-13
+# [2.0-rc.4] - 2022-02-15
+### BREAKING CHANGES
+- added MAX_INDEX, INDEXINFO_CACHETIME, NEWINDEX_CHECKTIME settings to default/sample web config file, copy to your config file
+- web config password for diskover user required to be hashed, create hash with password_hash.php and update your config file
 ### fixed
 - reduced login time when many indices
 - issue with not staying logged in and getting logged out before login time limit expires
 - issue with extra field value text on file view info page not wrapping when text string is very long
 - issue with extra field and object type not displaying correctly
+- changing chart/tree filter settings such as SIZE_FIELD in config did not change until browser cookies were cleared
 ### added
+- MAX_INDEX, INDEXINFO_CACHETIME, NEWINDEX_CHECKTIME to default/sample web config file, copy to your config file
 - password_hash.php - login password hash generator form
 - change password to settings page and warning if using default password
 - maxindex config setting to default/sample config, copy to your config