+++ diskover/CHANGELOG.md
@@ -1,12 +1,15 @@
 # Diskover v2 Community Edition Change Log
 
-# [2.0-rc.4] - 2022-02-11
+# [2.0-rc.4] - 2022-02-15
 ### fixed
 - issue with scanning in Windows
 - issue with setting domain to True in ownersgroups section in diskover config would case the scan to fail
 - UnicodeEncodeError exception when logging Unicode utf-8 file path warnings
 ### added
+- dir_depth, size_norecurs, size_du_norecurs, file_count_norecurs, dir_count_norecurs to ES index field mappings
+    - additional fields added to directory docs
 ### changed
+- hardlink files won’t be added to size_du (allocated size) of directories
 - set number of scan maxthreads when empty/blank (default) in config to number of cpu cores
 - indexing unrecognized Unicode utf-8 characters in file name or parent path, the characters are replaced with a ? character and file gets indexed with a warning log message
     - previously the file/directory was not indexed and just skipped with a warning message